serving/lambda/requirements.txt
# Lambda dependencies for API Gateway bridge
pydantic>=2.0.0
psycopg2-binary>=2.9.0
# boto3 is provided by the AWS Lambda Python runtime
